# Faster vec (un)pack