--- a/clash-prelude/clash-prelude.cabal
+++ b/clash-prelude/clash-prelude.cabal
@@ -424,3 +424,4 @@ benchmark benchmark-clash-prelude
                     BenchRAM
                     BenchSigned
                     BenchUnsigned
+                    BenchVector