Before:
```
benchmarking Vector/pack
time                 662.0 ns   (660.4 ns .. 663.4 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 659.9 ns   (658.5 ns .. 661.4 ns)
std dev              4.740 ns   (4.109 ns .. 5.578 ns)

benchmarking Vector/unpack
time                 1.197 μs   (1.196 μs .. 1.200 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.197 μs   (1.196 μs .. 1.201 μs)
std dev              6.806 ns   (2.929 ns .. 12.70 ns)

benchmarking BitVector/split# WORD_SIZE_IN_BITS
time                 52.77 ns   (52.58 ns .. 53.07 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 52.71 ns   (52.63 ns .. 52.83 ns)
std dev              312.5 ps   (209.9 ps .. 502.4 ps)

benchmarking BitVector/split# (3*WORD_SIZE_IN_BITS)
time                 61.04 ns   (60.98 ns .. 61.15 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 61.08 ns   (61.04 ns .. 61.18 ns)
std dev              219.1 ps   (132.5 ps .. 351.4 ps)
```
after:
```
benchmarking Vector/pack
time                 491.6 ns   (490.5 ns .. 492.9 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 491.6 ns   (490.9 ns .. 492.8 ns)
std dev              2.963 ns   (2.181 ns .. 3.994 ns)

benchmarking Vector/unpack
time                 796.2 ns   (795.5 ns .. 797.0 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 796.5 ns   (795.9 ns .. 797.5 ns)
std dev              2.521 ns   (1.787 ns .. 3.911 ns)

benchmarking BitVector/split# WORD_SIZE_IN_BITS
time                 39.09 ns   (39.08 ns .. 39.10 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 39.15 ns   (39.12 ns .. 39.22 ns)
std dev              134.2 ps   (70.88 ps .. 252.6 ps)

benchmarking BitVector/split# (3*WORD_SIZE_IN_BITS)
time                 59.10 ns   (59.09 ns .. 59.12 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 59.12 ns   (59.11 ns .. 59.15 ns)
std dev              68.60 ps   (45.69 ps .. 118.2 ps)
```